requirements.txt
tqdm
pytorch>=1.13.1
networkx
numpy
neo4j-python-driver
pandas
transformers
datasets
nltk
python-louvain
openpyxl
beautifulsoup4